requirements.txt
fastapi
uvicorn
python-multipart
openai
python-dotenv
requests
httpx
aiofiles
gunicorn
celery[redis]
redis